safari-534.53-branch/LayoutTests/http/tests/security/sandboxed-iframe-form-top.html

<html>
<head>
<script>
if (window.layoutTestController)
    window.layoutTestController.dumpAsText();
</script>
</head>
<body>
<p>This tests passes if the sandboxed frame cannot navigate the top frame.</p>
<p>PASS</p>
<iframe sandbox="allow-forms allow-scripts"
        src="resources/sandboxed-iframe-form-top.html">
</iframe>
</body>
</html>
